move provable intf
