+++ src/index.ts
@@ -1,4 +1,4 @@
-export type { ProvablePure } from './snarky.js';
+export type { ProvablePure } from './lib/provable-types/provable-intf.js';
 export { Ledger } from './snarky.js';
 export { Field, Bool, Group, Scalar } from './lib/core.js';
 export {
