+++ src/lib/circuit.ts
@@ -1,10 +1,11 @@
 import 'reflect-metadata';
-import { ProvablePure, Snarky } from '../snarky.js';
+import { Snarky } from '../snarky.js';
 import { MlFieldArray, MlFieldConstArray } from './ml/fields.js';
 import { withThreadPool } from '../snarky.js';
 import { Provable } from './provable.js';
 import { snarkContext, gatesFromJson } from './provable-context.js';
 import { prettifyStacktrace, prettifyStacktracePromise } from './errors.js';
+import { ProvablePure } from './provable-types/provable-intf.js';
 
 // external API
 export { public_, circuitMain, Circuit, Keypair, Proof, VerificationKey };
