+++ src/lib/provable.ts
@@ -5,7 +5,7 @@
  */
 import { Bool } from './bool.js';
 import { Field } from './field.js';
-import { Provable as Provable_ } from '../snarky.js';
+import type { Provable as Provable_ } from './provable-types/provable-intf.js';
 import type { FlexibleProvable, ProvableExtended } from './circuit-value.js';
 import { Context } from './global-context.js';
 import {
